Radiation evaluation for conifets and hardwoods - compiles not debugged yet.
Logic of radiation calculations:
(See algorithms.cc for examples of use)

Friend function of tree EvaluateRadiationForTree takes the tree and the functor
WrapRadiationEvaluations as argument.

WrapRadiationEvaluations takes care of handling conifers and
and hardwoods correctly in calculation of radiation (conifer and
hardwood functors wrapped into).

WrapRadiationEvaluations 'runs' ForEach using functor
EvaluateRadiationForXXTreeSegment (XX = Cf or Hw), which in turn
'runs' Accumulate with functor ShadingEffectOfCfTreeSegment or
ShadingEffectOfLeaf to check whether all other segments or leaves
are shading the subject segment or leaves of the subject segment.

Conifer and hardwood versions of beamShading (cylinderBeamShading,
ellipsisBeamShading) do the actual checking of ray interception.