stdafx.h  is an  empty header  file for  Unix systems  that  should be included in every source file (suffix cc, cpp, cxx etc.). Furthermore, the line
	#include <stdafx.h>

must  be  the very  first  sentence  in  the source  file.   Including
stdafx.h is not necessary in header files (suffix h).

This requirement on Unix is  due to Microsoft Visual C++ compiler. The
MS VC++  manual says  every source file  must include stdafx.h  in the
beginning  of the  file.   Unfortunately  it is  not  possible to  use
conditional  compiling.   MS  VC++  seems  to override  all  rules  of
conditional  compiling and  checks  that the  file  stdafx.h is  truly
included.