This  commit synchronizes the  files between  Mika's work  for Windows machines the work done for Unix/Linux. The framework for conferous and hardwood species contains now everything necessary to implement growth of indivudual tree species:
1. Radiation calculation
2. Photosynthesis and respiration
3  New structural growth (diameter growth, vigour index etc.)
4. Allocation of photosynthates (zbrent)
5. Senescence.

The implementations don't always follow  the idea of using the generic
algorithms (to traverse the tree)  and functors, there are even global
variables  used. So  there is  some cleaning  up to  do  including the
removal of warnings in the compiling. But Eero should able to start ro
work with wihite birch and the changes can be done behind the scenecs.