The  work organizing  header files  allowing single  <Lignum.h> header file   in   the  Lignum   API   (application  programming   interface) continues. When working with the header files in stl-lignum please try to quickly  organize included header  files so that the  system header files come first (i.e. files not related to lignum and c++adt).
I have some  mysterious problem in algorithms.cc. The  ordering of the
header files must be:

	#include <MyTreeSegment.h>
	#include <Lignum.h>
	#include <MyBud.h>
	#include <Shading.h>

the ordering:

	#include <Lignum.h>
	#include <MyTreeSegment.h>
	#include <MyBud.h>
	#include <Shading.h>

does not compile. Also, as part  of the Lignum API, Shading.h could be
included in Lignum.h. Simirlarly  other radiation related header files
and header files related to growth of an abstract tree.