Header  files   implementing  L-systems  in  LIGNUM   are  in  include directory.  Most notably: LSystem.h provides programming interface for rewriting.  lsystem.h implements  the rewriting and lengine.h includes necessary  system   files  for  LSystem.h   and  lsystem.h.   Turtle.h implements the turtle movements (included in lengine.h).
LSystem.h and lsystem.h are NOT  and include header files that are NOT
wrapped with ifdef directives  because LSystem.h and lsystem.h must be
able to be included several  times, once for each species.  To include
the  files several  times namespace  must  be used  in L-system  file.
LSystem.h includes lsystem.h.

In short the  files of interest for a user  are lengine.h (required by
LSystem.h  and lsystem.h),  LSystem.h providing  programming interface
for  rewriting and  lsystem.h implementing  the rewriting.  See README
file for further details.