Added functor CreateLeaves that creates leaves to a broadleaved trees. Usage: 	vector<PositionVector> v; 	AccumulateDown(tree,v, 		       AppendSequence<vector<PositionVector> >(),                        CreateLeaves<MyHwTreeSegment,MyHwBud>());
The algorithm is based on LGMstatus field of a Bud. If status is 1 the
direction  vector  of  the  bud  is  added to  'v'  and  passed  down.
Branching  point appends  incoming  vectors and  passes the  resulting
vector downwards  to the  segment below that  creates the  leaves. The
vector is  cleared by the segment  so that the leaves  are not created
twice. AppendSequence<S>(s1,s2) is a general purpose functor to append
the the contents of s1 to the end of s2. s2 can be an empty sequence.