TMatrix3D implements 3D  matrix where indexing goes in  C style. Given TMatrix3D<T> m3, m[x][y][z]  gives the element T in  a cell indexed by x,y and  z.  The implementation  is inspired by Breymann  (Chapter 9). At first user may wonder where are the indexing operators but they are provided by inheritance.  Index  checking indexing is provided by at() method.
With the class TMatrix3D<T> I want  to demonstrate the use of STL, its
classes  and algorithms,  and the  importance of  clear understandable
implementations at  the possible cost of small  runtime efficiency.  A
correct program is  more important than a fast  program.  The approach
here can  be easily expanded  to N dimensional matrices  (exercise for
the reader).