Sample functor DeleteDeadBranches  to delete axes (branches)  in the tree. Axis: If an axis can  be deleted mark  it dead.  BranchingPoint: Delete  the dead axes and erase them from  the the list.  TreeSegment: Collect foliage mass, the  criteria to delete axes.  Algorithm  ensures that 'delete' is not applied to main axis.
There are a  couple of caveats in deleting  axes and branching points
in a  tree this way.   Due to the  fact that the  main axis is  not a
pointer, we cannot  simply apply 'delete' on axes.   That would cause
an error if delete is called on the main axis.  Also if explicit call
to ~Axis() is  used this can result problems  on program exit: (main)
axis may already have been  destroyed inclusive the list of branching
points. This means  that the list does not  have any methods availabe
either! Naturally  call to any list  method in such  case will result
runtime error. Note  also that destruction and erasing  the axes must
be done at the same  place (BranchingPoint). Once deleted there is no
way to check  afterwards (in some other place) if  an axis is deleted
or not; using/checking deleted pointer results unspecified behaviour.
Sometimes I just love C++...