Let's try to be careful with these:
1. Constructor: lr  (lenght radius  ratio) should be  [0:1]. I  try to
have parameters that consistently  are multiplied. (not both divisions
and multiplications,  too much to remember). Intial  foliage mass (Wf)
is simply af*Sa where Sa is the surface area of the segment and af the
parameter.  Foliage height  Hf is  simply nl*sin(na)  where nl  is the
needle length and na needle angle.

2.GetValue:  Af is  foliage area  computed as  sf*Wf where  Wf  is the
foliage  mass  and  sf  parameter  (c.f. parameter  af  and  don't  be
confused). Radius  to foliage limit,  Rf, is now  R+Hf where R  is the
segment radius and Hf is the  foliage height (so when the segment adds
diameter the foliage is not sunken in the segment).