The pairwise comparison of coniferous segments to compute Qin and Qabs is now  two nested ForEach  calls. Change the Accumulate  when chasing subtle bugs in the computations.
Now  the functors  for ForEach  should  not generate  calls to  vector
constructors (K  and shading vector  S) but still  the is lost  to the
list traversing in pairwise comparison of segments.

And the bug: I think I found it. In ShadingEffectOfCfTreeSegment it is
O.K not to compare  the segment to itself. But I could  not find if it
is  checked whether  the shading  segment  is higher  than the  shaded
segment. The comparison  results a hit in to  all directions (I think)
but we do not look downwards. Added the check.

Now  the Qin  seems to  vary  reasonably. Previously  Qin was  similar
everywhere in the tree and gradually diminishing, even the top segment
was  very   shaded.  And   the  result  was   sudden  death   (due  to
f(ip)-function).