Changed the example of the usage of Gauss normally distributed random number generator.
It is not quite obvious how to generate different sequences of
normlly  distributed random numbers. The sequence of random
numbers is initialized (call to init()) with any negative seed
and different negative seeds will initialize the sequence in different ways.

The subsequent calls to the functor with some positive seed will
produce the numbers from the sequence. Different positive seeds
do not reinitialize the seqence.  To be more obvious, the functor
should not have any arguments at all. Now the functor has number 1 as
a default argument.

The example is specifically for N(0,1) printing probabilities for
[-1.96,1.96] = 95% and [-1,1] = 68%.

Usage: gauss init seed mean stddev sequence_length