Changed the  example of the  usage of Uniform  uniformally distributed random number generator.
It is not quite obvious in Uniform how to generate different sequences
of  uniformally distributed  random numbers.   The sequence  of random
numbers is  initialized (call  to init()) with  any negative  seed and
different  negative seeds  will initialize  the sequence  in different
ways.

The  subsequent calls  to the  functor  with some  positive seed  will
produce the numbers from the sequence. Different positive seeds do not
reinitialize the seqence.  To be  more obvious, the functor should not
have any arguments  at all. Now the functor has number  1 as a default
seed.

Usage: uniform init seed