Changed the LGMAD symbols As0 and age to LGAAs0 and LGAage respectively. Risto and I are doing the c�ean up of the symbols in LGMSymbols.h.
1. There shall be only one symbol for one purpose.
2. Uncommented symbols will disappear (after a couple of warnings)
3. Do not confuse between LGMAD and LGMPD symbols
4. Do not use symbol names as variable names.

All four items are a bit broken now. Especially, age was used
quite a bit as a variable name in Graphics.

It is O.K to have structure member with a symbol name

struct A{
  int age; <-- O.K., no confusion
};

But it is NOT O.K to have simple variable matching th symbol name.

void f()
{
   int age = 0;  <-- O.K. for the compiler, hides the symbol!!
   .....
   GetValue(ts,age); <-- O.K. for the compiler,
                         but you won't get the
                         age of the segment!!!!!
}