Ellipse::setArea(A,scaling_point)  had a bug:  The algorithm  was like the setArea(A) where  the scaling point is the  center of the ellipse. It did not  take into account that the center  point will translate in space due to the scaling operation.
Added a  quick fix  for the SugarMaple  project purposes: 1.   Set the
scaling point as  the new center. 2.  Scale 3.  Move the scaling point
to    the    new    center    point   after    scaling:

      center = base + x1u()*semimajoraxis.

This  works if  the end  of  the petiole  (i.e the  scaling point)  is
attached  to the  ellipse  leaf,  so that  it  is at  the  end of  the
semimajoraxis.  x1u()  is  then  the   direction  to  go  to  and  the
semimajoraxis after scaling tells how much.

From the Ellipse  point of view we should do  the scaling properly, so
that the scaling is truly generic (obviously the scaling_point must be
inside the Ellipse, no need to be more generic than that).