The  configuration  of  the  mask  can   be  given  in  a  file  as  a ParametricCurve.   The  mask  is   a  function  of  inclination.   The inclinations run  from 0 (horizon) to  90 (zenith) and  the mask value can be [0:100] as before (0 is full light).  The name of the mask file is  given in  the file  defining the  sky (inclinations,  azimuths and radiant intensity).  Typically the file defining the sky looks like:
	     #Sectors
	     18    #incl
	     10    #azim
	     1450  #Radiation MJ
	     #Gap data follows
	     GapData.fun

This is the recommended way.   The algorithm in readMask() still copes
with the old way  too.  It checks the type of the  first token. If the
token is an identifier, it is assumed to be the name of the mask file.
If the identifier is a number,  then the old style listing of the mask
is assumed to follow.