Implemented method prune that removes dead axes from the tree, updates and  synchronizes  the  L-system   and  the  tree.  The  interface  is overloaded:
	lsystem.prune(tree);
        lsystem.prune(tree,predicate,init);

The first one uses DeleteDeadBranches to delete axes. The second one uses
used defined predicate function operator to delete the axes. In fact,

        lsystem.prune(tree);
and
	double foliage = 0.0
        lsystem.prune(tree,DeleteDeadBranches<TS,BUD>(),foliage);

are equivalent. Any predicate 'p'  with some initial value 'init' will
do as long as

        AccumulateDown(tree,init,p)

can be compiled. Both pruning methods have longer forms to interchange
data with the L-system and the tree, for example:

	lsystem.prune(tree,1,Name);
        lsystem.prune(tree,predicate,init,1,Name);

where   the  attribute   with  the   name  'Name'   will   be  updated
(c.f. lstringToLignum and lignum2Lstring).