Added method readAllMasks that configures the sky from  the file with a sequence of mask  files.  The  file format  is:   Format          |  For example   INCL            |  9   AZIM            |  9   RAD_PLANE       |  1200   ITER1 MaskFile1 |  5   Mask1.fun   ITER2 MaskFile2 |  10  Mask2.fun   ITERN MaskFileN |  100 Mask3.fun
After  giving the  size and  the radiant intensity,  N mask  files may
follow.   In  the example  above  the  iterations  [0:5] will  use
Mask1.fun,   the   iterations   [6:10]   Mask2.fun   and the iterations
[11:100] Mask3.fun.

Added method configure(int i, bool verbose) that will configure
the sky according to the mask matching the 'i':th iteration.

The readAllMasks is used by  InitializeTree, it also calls
configure(0,verbose) to set the first mask.

It is then up to the user to call configure(iter,verbose) to change the mask
in the growth  loop.  If verbose = true, configure will produce vebose
output, the default value is false.