ReDrawWindow now displays the help string at the bottom of the window. It also maintains its relative position to the camera, thus it remains still (i.e. appears to to be) while the camera moves.
Implemented Zoom(int dir) that moves the camera and the point of focus
forward (dir = 1) or backwards (dir = -1). Implemented
Translate(PositionVector& d) that moves the camera and the point of focus
up, down, left and right.

StartAnimation now assumes that the user wnats to see the current ShowTree
tree a little closer. Zoom and Translate may have changed the point
of focus from the ShowTree. To make the requested rotation meaningful
reset the focal point the ShowTree tree.

Added the new keys for Zoom and Translation to Keypress().