Added  method getRoute(v,p0,dir)  that  generalizes the  trace of  the light beam in the voxel space. That is, the user given Point p0 is the start point of the ray and PositionVector dir its direction. The route from voxel box to another is stored in the vector<VoxelMovement> v.
The free  p0 requiers the  lengthy initialization that checks  all the
faces of  the box for the  light beam intersection  of before entering
the  incremental phase  of the  fast voxel  traversal  algorithm.  See
description how  to check if  a ray intersects  a plane in  the source
file.

I  have tested  getRoute with  basic cases  (direction up,  down, left
right etc.) and  with a bit more complicated  ones. The initialization
seems to  be O.K.   Also note that  the voxel traversal  algorithm can
handle cases where  the light beam hits a box corner  or where two box
walls (planes) intersect.