Voxel Object book keeper takes care of the problem of duplicate (or more!) computing of extinction when the segment is split into several parts. Each object has a tag number given by the book keeper. If two objects denote the same segment, they have the same  tag.
Book keeper also has a boolean vector denoting if a ray has hit
foliage of segment represented by voxel object with a tag number n.
Before computing the extinction, book keeper can check if v[n]==true.
If so, the extinction is already computed (Recall CylindeBeamShading
returns the full path length of the ray in the foliage, not the path length
of the ray in the foliage in the box).