Added gravelius order for the BranchingPoint. Also edited SetValue/GetValue to recognize the LGMomega symbol. For the cowberry, to pass the information from terminating buds to lateral buds, we need to store temprarily the information in the right branching points (determined by the gravelius order).
So adding Gravelius order to branching point is O.K., it does not break any code,
and might be uselful in general. The gravelius order is set in LEngine,
otherwise user must set it explicitely.

Typically, we need to store information in branching points when we
cannot directly, in a single traversal, pass the information to the
tree compartments where it is needed. Currently TreePhysiologyVigorIndex
is such and now in Cowberry terminating buds affect the development
of lateral buds. We cannot directly pass the information between buds,
but first AccumulateDown the information from terminating buds
to the right BranchingPoints and then PropagetUp the information
from the BranchingPoints to buds where the information is needed.
Added 'type' that keeps the type of the terminating bud of the same axis
where the BranchingPoint is.

The problem in the longer run might be that this type of two-way
passing of information  for various purposes little by little
expands the BranchingPoint unnecessarily. We may need to abstract
the BranchingPoint as TreeSegments and Buds, but this requires
careful thinking. BranchingPoint was thought to be quite stable
just being a container for  the list of axes.