Added the possibility for calling external programs. Programs are called by writing the commands to the command line found in the lower part of the window above the text output box, and clicking the 'Run'-button.
External programs can also be stopped with 'Kill program'-button, but this
might still be a bit buggy.

Also the working directory can be changed from 'Options'->'Change working
directory' as some external programs might need files etc. to be found
in the current working directory. Note that changing the working directory
changes the working directory of the whole workbench, not only the external
programs. This affects the loading of textures as by default they are
loaded relatively from the current working directory of the workbench.
The file names can though be changed to absolute values from the visualization
options dialog.