GUI-class for the XML-viewer.
Currently the viewer is a separate window which can be used as a stand alone
application if there is a need. There isn't yet a .ui-file for the viewer,
but I shall create one if there is a need the have more GUI-components
for controlling the viewer (which is more than likely).