Thread class for loading the BSP-tree outside the main thread. Should improve the general usability of the workbench as the GUI doesn't get blocked any longer while loading trees from XML viewer to visualization.
However, loading new files to XML viewer still blocks the GUI, but that can
be done in a it's own thread also.

It would also be nice to display some kind of progress bar beside the XML
viewer telling how much of the BSP-tree is constructed, but that can quite
tricky to do as the actual size of the BSP ain't known until it's fully
constructed. However, the bar doesn't have to be 100% precise.