Added specific leaf area as a CfTreeSegment attribute. Default value is the tree level parameter set in the constructor. Other implementations may follow in actual tree species (see e.g. Scots pine).
Implemented SetValue and GetValue for LGAsf. Also changed LGAAf
to return LGAAf = GetValue(ts,LGAsf)*GetValue(ts,LGAWf).

This is due to the computations of light climate in the voxel space
implemented in VoxelSpaceRadiationI.h, more precisly calculateAbsorbedCfRadiation,
that requiers sfk = GetValue(ts,LGAsf), i,e, sf as a segment parameter.
This in turn is the consequence of experiments in Lig-Crobas where Scots pine
has LGAsf as a function of light for each segment.

As the initial value is the tree level LGPsf, applications assuming
it when quering GetValue(ts,LGAAf) behave as before.