PairwiseVoxelSpaceRadiation can now compute radiation attenuation for both conifers and hardwoods. To compute for conifers use it as
	PairwiseVoxelSpaceRadiation<Segment,Bud> r(vs,K,1.0);
	ForEach(conifer,r);

For hardwood species use it as

	PairwiseVoxelSpaceRadiation<Segment,Bud,Ellipse> r(vs,K,1.0);
	ForEach(harwood,r);

Why like this? PairwiseVoxelSpaceRadiation uses the voxel space to
follow the light beam. Each voxel carries information about the
shading elements (segments, leaves) and these elements are checked
for ray-shape intersection. The information of the shading elements
is 'wrapped' inside VoxelObjects and each voxel object can calculate
the radiation attenuation (wheter it represents a segment cylinder or leaf).
So for the conifers and broadleaved trees the Qin part is the same.
They differ in the Qabs part.

I have not tested the implementation for hardwood yet, but ScotsPine
in Lig-Crobas functions as before.

By (re)implementing PairwiseVoxelSpaceRadiation and some changes
in CfTreeSegment (LGAsf now attribute) we can now set up a project
for Juan called OptimalTree and start implementing necessary
algorithms.