InsertVoxelObjects now wraps the insertion of voxel objects for both conifers and hardwood. Usage:
  InsertVoxelObjects<CfSegment,CfBud> dmp(voxelspace,dir,dist,beam_start,num_parts);
  ForEach(conifer,dmp);

  InsertVoxelObjects<HwSegment,HfBud,Shape> dmp(voxelspace,dir,dist,beam_start,num_parts);
  ForEach(hardwood,dmp);


Where dir and dist define the direction and distance where to insert the tree compartment
(when using copies of the original tree as forest).

For conifers num_parts means the number of points on the segment cylinder checked
when inserting the segment into space. For hardwood num_parts is the number of
points on the shape boundary checked. The mid point of the shape is always checked.
Note: currently Ellipse is implemented.

For conifers beam_start is the relative start point [0:1] of the light beam
on the segment (we need the information to check if the start of the
light beam is on the shading segment-->comparing the same segments).
For hardwood it is not implemented (we assumme the midpoint of the leaf).

To compute Qabs using voxel objects ("pairwise comparison") use

   PairwiseVoxelSpaceRadiation<CfSegment,CfBud>(vs,K,beam_start) qabs;
   ForEach(conifer,qabs);

   PairwiseVoxelSpaceRadiation<HwSegment,HwBud,Shape>(vs) qabs;
   ForEach(hardwood,qabs);

beam_start is the relative start point [0:1] of the light beam on the segment.