HwVoxelObject  takes  now  as  the  leaf itself  in  constructor.   In OptimalTree the leaf orientation is changed frequently, and the change is immediately evident  as the leaf pointer in  the VoxelObject points to the leaf itself. This evades calls to InsertVoxelObjects.
Naturally, in getRoute() if the leaf itself is used then its Ellipse
is used.

This may sound as a "cake on a cake" because the geometric information
is also  stored, but the thing  is that so  far a forest in  LIGNUM is
created by  using copies of  a single tree  and the leaf  ellipses are
inserted to specified  positions. If the original leaf  were used, the
locations  would be lost,  because the  leaf pointers  in VoxelObjects
point the original leaf.