parseBranchingPointAttributes  had  a  bug.   It only  set  the  first attribute LGAMaxD and then jumped  out of the loop. Change the looping so that all attributes are set.
parseCfTreesegmentAttributes  set the LGAVf  to LGAstarm.  Removed the
erroneous  SetValue,  but left  the  parsing  of  LGAVf (for  backward
compatibility).

It seems that reading and writing of XML files is correct (copy-tree
in OptimalTree reads an XML file and writes it to another. The files
are the same except numerical accuracy).