Added ResetQinQabs that  sets LGAQin and LGAQabs to  0.0 in a segment.
NOTE  1: this must  be called  before any  attempt to  calculate Qabs,
because  I will  remove  the  resetting of  these  two variables  from
PairwiseVoxelSpaceRadiation      in      VoxelSpaceRadiationI.h     in
stl-voxelspace.   I must do  this, because  the implementation  of the
forest  consisting   of  individual  trees   claculates  the  absorbed
ratiation in two phases: 1) Green  voxel for the shading of the forest
and 2) PairwiseVoxelSpaceRadiation to calculate self shading.

NOTE 2:  I will add the  call to ResetQinQabs in  Lig-Crobas, but this
inevitable     change     has     changed     the     behaviour     of
PairwiseVoxelSpaceRadiation.  So the change  in Lig.-Crobas  is needed
anywhere, where PairwiseVoxelSpaceradiation is used.