Added functor (TreeFunctor.h) MoveHwTree that moves also leaves of a hardwood tree. Implemented it with methods in BroadLeaf and Triangle, that move the corners of the triangle (and petiole). Does not work yeat with shape Ellipse. Needs grooming.
TranslateLeaf(BroadLeaf<SHAPE>& bl, const PositionVector& t) of BroadleafI.h
did not work:  SetCenterPoint(bl, pnt + t_point); seemed not to work
(with Triangle). There is something odd in CenterPoint of Triangle.
It did not update (in a meaningful way) either, when the corner points
of the triangle were set.