The content of this commit is identical to the previous one, only that there is a crucial typing error in the log. This is the correct one:
I added functors ShadingEffectOfWoodyPartsSelfConifer and
ShadingEffectOfWoodyPartsConifer. They can and should be used to evaluate
transmission when only the woody parts of conifer trees are considered.
They differ from ShadingEffectOfWoodyPartsSelf and
ShadingEffectOfWoodyParts in that they consider only those shading
Segments that do NOT carry needles. This is to prevent double accounting
of woody parts of segments with needles (their woody parts are accounted
for in their transmission).

ShadingEffectOfWoodyPartsSelf and ShadingEffectOfWoodyParts can be technically
used for conifer shading but they result (probably) in double
accounting.