I added to     void DumpCfTree(VoxelSpace &s, Tree<TS, BUD> &tree,int num_parts, bool wood)
updating values in VoxelBoxes after dumping the tree.
It essentially evaluates STARmean value with the aid of STARsum and weighting
factor (needle area).

I changed in VoxelBox::updateValues()
that it is checked from  number_of_segments_real
if there are anything to calculate for STARmean.
number_of_segments (integer) is a bit problematic when segments are put in VoxelBoxes
in parts.

It seems that VoxelSpace has many layers of programming efforts. It may be necessary
to check that everything is done consistently in the whole library. E.g. using
VoxelBox::updateValues() - and not just updating some necessary variables in
the very application.