Type parameterized LGMSplit: the actual "cut" function/functor can be given as a template type: "LGMSplit<TS,BUD,F=CutSegment> split". Default type (i.e. functor) is CutSegment that splits segments into equal lengths. The functor operator signature in CutSegment is as follows:
		bool operator()(list<TreeCompartment<TS,BUD>*>& tc_ls,
				typename list<TreeCompartment<TS,BUD>*>::iterator iter,
				METER l_new, METER l_min);

tc_ls: is the list of tree compartments (i.e. axis)
iter: is the pointer (list iterator) to the current tree compartment
      (NOT necessary a tree segment) in the list
l_new: is the length of the new segment
l_min: is the minimum allowed length of the segment.

The default cut operator creates a new segment, sets the length of the
new segment to l_new and subtracts l_new from the current segment.
Finally the new segment and a new branching point are inserted in front
of the current segment.

In most cases one can use CutSegment to split the segment. Naturally, if
application specific variables need update, that must be done after the split.

Note that for the application specific updates the new segment is simply two list elements away:

		iter2 = iter //iter2 points to current tree segment
		advance(iter,-2)//iter2 points to the new segment

See SplitFungiSegment in Heterobasidion as an example of an application
specific tree segment split functor.